Move truncateString util into @uppy/dashboard.
