+++ packages/@uppy/dashboard/src/truncateString.js
