+++ packages/@uppy/dashboard/src/truncateString.test.js
